% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSilvaRef.R
\name{buildSilvaRef}
\alias{buildSilvaRef}
\title{Create sequence and taxonomy file for microbiome analysis}
\usage{
buildSilvaRef(
  fin,
  ftax,
  fout_fasta,
  fout_taxonomy,
  include.species = TRUE,
  compress = FALSE,
  n_euk = 100
)
}
\arguments{
\item{fin}{Path to the input SILVA fasta file (RNA sequences).}

\item{ftax}{Path to the SILVA taxonomy map file with valid taxon strings and levels.}

\item{fout_fasta}{Path to the output fasta file.}

\item{fout_taxonomy}{Path to the output taxonomy TSV file.}

\item{include.species}{Logical; whether to include species names in taxonomy.}

\item{compress}{Logical; whether to compress output sequences using gzip.}

\item{n_euk}{Integer; number of random eukaryotic sequences to include in the final output.}
}
\value{
Writes a fasta and a taxonomy file to disk. Returns nothing.
}
\description{
Reads a SILVA fasta (tested on NR SSU set) and corresponding taxonomic map file, extracts
annotations, filters them, and outputs a refined fasta and taxonomy file. Optionally
includes a set number of eukaryotic sequences. The matchGenera function is adapted
from the dada2 R package.
}
\examples{
\dontrun{
buildSilvaRef("SILVA.fasta", "SILVA_taxonomy.tsv",
                     "output.fasta", "output_taxonomy.tsv",
                     include.species = TRUE, compress = FALSE,
                     n_euk = 100)
}
}
