% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildUniteRef.R
\name{buildUniteRef}
\alias{buildUniteRef}
\title{Create formatted reference fasta and taxonomy files from a UNITE general release}
\usage{
buildUniteRef(
  fin,
  fout_fasta,
  fout_taxonomy,
  include.species = TRUE,
  compress = FALSE
)
}
\arguments{
\item{fin}{Path to the input UNITE FASTA file.}

\item{fout_fasta}{Path for the output FASTA file.}

\item{fout_taxonomy}{Path for the output taxonomy TSV file.}

\item{include.species}{Logical; if TRUE, includes validated species names.}

\item{compress}{Logical; if TRUE, the output FASTA file will be gzip compressed.}
}
\value{
Function writes two files to disk, but does not return values.
}
\description{
The buildUniteRef function processes a UNITE reference fasta file by extracting and
filtering taxonomic annotations. It validates species names, handles problematic labels
and produces two separate output files: a reference fasta sequence file and a
tab-separated file containing reference ID and filtered taxonomic classification.
}
\examples{
\dontrun{
buildUniteRef("sh_general_release_dynamic_10.05.2021.fasta",
              "unite_ref.fasta", "unite_taxonomy.tsv",
              include.species = TRUE, compress = FALSE)
}
}
